Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
mysql-connector-python==9.4.0
bcrypt==4.3.0
python-dateutil==2.9.0.post0
gunicorn==21.2.0
python-dotenv==1.0.0